Adding the ability to upload backups to s3
It works on the backup.backup_s3_bucket setting in the config, which
translates to a settings.BACKUP_S3_BUCKET django setting.

If this setting is None, then it silently assumes that you don't want to
upload to S3.

It won't make a backup if it can't find the S3 bucket to put it in
(questionable if this is the right thing to do)